Fix for HTTP 204 returned with content data (Exception stack) during resource throw exception in HTTP2  (#526)
When the resources return 204, sometimes they throw an exception. For whatever reason, those exception will become the attached data in the response. The reason we do this PR is to remove the data because for HTTP status 204, it literally means no data attached (No content). It was discovered by the condition where we exercise HTTP2 migration.