Shade Netty.
Per internal discussions, we'd like to shade Netty usages in rest.li,
so this change relocates io.netty to com.linkedin.pegasus.io.netty.

A new module named li-netty is introduced which produces an artifact
in the 'shadow' configuration that includes the Netty code relocated,
and all references to io.netty in rest.li are changed.