Fix bottlenecks in DataSchemaParser (#616)
* Fix bottlenecks in DataSchemaParser

1) Remove usage of TreeMap. Using a TreeMap with a DataSchema key is
very expensive. Every single comparison was calling toString() which
forces the both schemas to be serialized to JSON, then compared.

The point of this TreeMap was just to have a deterministic
ordering, when the source file ordering has changed. To maintain this
behavior we can just sort the input upfront, and maintain a
LinkedHashMap for consistent ordering.

2) Replace binding() call in loop with existingDataSchema(). The
resolver instances use  AbstractMultiFormatDataSchemaResolver, which
has a very expensive implementation of binding(). It creates a new
aggreagted map on every invocation. So calling this once for a single
lookup is actually more expensive than a linear search would be.
Replaced this with existingDataSchema() which is an equivalent much
cheaper operation.

Both of these optimizations combined, reduce client generation
code-generation from 2 minutes to 20 seconds on a large codebase.

* Bump version and update changelogs