Throw an exception when an entity-level action would be invoked without an id.
Entity-level actions are performed on a specific resource, and so the id value
in the request is mandatory. Today, if the client forgets to call .id(...) to
supply a value the request can be built and sent to the server where it will
cause an error with a mysterious error message that implies that there is no
action of that name, i.e.
"POST operation named myAction not supported on resource '...' URI: '...'"

Here we fix that so that
a) The problem is caught on the client side during the RequestBuilder.build()
   method call so that the request cannot be sent to the server.
b) A sensible error message is given where the stack trace's line number will
   lead the developer straight to the root of the problem:
   "Missing or null id value; we cannot invoke this entity-level action."