Revert DefaultDocumentationRequestHandler to be blocking again (#824)
Back in January, we changed the blocking DDRH to be non-blocking (fail requests sent during initialization)
to solve an edge-case issue of request thread pool exhaustion. This started causing a lot more complaints,
as many more use cases were now seeing errors for OPTIONS calls that would otherwise be successful.

This change brings back the old blocking behavior as default, and adds NonBlockingDocumentationRequestHandler
as a subclass of the DDRH which users can configure if they see thread pool exhaustion in their use case.