Add RestLiTraceInfo to the request context for both client and server (#846)
Defines a unique identifier for a given endpoint called a resource method identifier (e.g. "parent/{}/sub/{}:finder:search").
Exposes this identifier in both the client Request and in the server ResourceMethodDescriptor.
Introduces a utility called ResourceMethodIdentifierGenerator which constructs this RMI.
Adds this identifier to the request context on client and server as Rest.li trace info.