Fix conversion of R2 headers to Netty headers in requests.
If the headers have a null value it causes an NPE from Netty. This
treats a null header as empty String. Additionally, this fixes the
exception handling error that this behavior exposed. An NPE or any
other exception was causing the request to wait until timeout rather
than handling the error, due to the `ChannelPromise` not being passed
to write(), which creates a new promise that does not notify
listeners to the original one.