Improve synchronization on the R2 `RequestContext` local attributes.
It is possible to cause a `ConcurrentModificationException` in the
`RequestContext` because we synchronize it by using
`Collections.synchronizedMap()`, which only synchronizes individual
reads/writes. When making a `RequestContext` copy we must iterate over
the collection, allowing other threads to interrupt it by modifying the
map during the operation.

This fix synchronizes the whole copy operation as well. The default
mutex used for synchronizing in a `SynchronizedMap` is `this`, so
synchronizing on the map itself guards against modification.