Downgrade major version to 29 (#931)
Bumping the major version to 30 is technically the correct thing to do from a
semver perspective when a public API is removed. However, it also has a huge
impact that outweighs the API breakage in terms of propogating the change to
all consumers. We feel like this tradeoff is acceptable.