Use Node instead of D2Node and D2URIMap instead of NodeMap for xDS flow (#944)
* Use Node(Map) instead of D2node(Map)

The initial intent of using D2Node was to pre-deserialize the JSON, but because
there is no way for the rest.li client to actually leverage this, it simply
re-serializes the JSON then deserializes it again. This causes memory pressure
along with validation errors as numerical typesa get mungled.

* Leverage pre-deserialize JSON struct instead of switching over to raw JSON data

Additionally introduce a D2URI type which has a proper schema instead of being a
Struct. This will significantly decrease the resource size sent by the observer
and significantly decrease the cost of deserialization on the client.

* Bump minor version

* Comment and simplify D2URI

* Meet proto standard of snake_case naming

* Update comment on D2URI

* Simplify D2URI even further

* Use the raw JSON Node type instead of D2Node

This is causing all sorts of validation issues due to integers being interpreted
as floats and vice versa

* Fix some compile errors

* Update changelog

* Fix test failing due to missing property

* Update changelog