Default to pick_first policy with shuffled addresses
This offers an improvement over the default "pick_first" policy which always
picks the first host returned by getaddrinfo, which may not always be random
(see linked documentation in `XdsChannelFactory`). The default policy used to be
"round_robin" but this makes the load balancer eagerly create a connection to
each xDS server, instead of only the one it is talking to. This is starting to
cause significant load on the observer instances.