Implement new DualReadMonitor for UriProperties
This new monitor should be as expensive as the previous one since the previous
one called the `.equals` method on `UriProperties`, which compares every single
URI. It now produces a similarity metric which represents the fraction of
matching hosts present in the ZK response and the observer response. The Cluster
and Service monitors remain unchanged.