Respect `startPublishing` call by always re-notifying watcher in XdsClientImpl
When switching from the backup to the primary store (implemented in the
XdsClientImpl), the data in the primary store is never replayed. This is
different from the backup store behavior which respects the invocation of
startPublishing and replays the contents of the store. This means that if the
contents of the backup store are different from the contents of the primary
store, and the client switches from the backup to the primary, the client will
only see the backup values and not the primary values.