Bump gRPC version to fix IPv6AwarePickFirstLoadBalancer
The signature of `LoadBalancer#acceptResolvedAddresses` was changed from
returning a boolean to a `io.grpc.Status`. This meant invoking
`IPv6AwarePickFirstLoadBalancer#acceptResolvedAddresses` was invoking the
default definition in the parent abstract class instead of the actual overridden
definition in `IPv6AwarePickFirstLoadBalancer`. Strangely, it did not cause any
runtime exceptions such as NoSuchMethodError.